Updated tests

+++ src/test/java/test/producer/ProductCreatorTestImpl.java
@@ -3,9 +3,6 @@ package test.producer;
 import ru.fd.api.service.entity.Product;
 import ru.fd.api.service.entity.ProductDefaultImpl;
 
-import static org.testng.Assert.assertFalse;
-import static org.testng.Assert.assertNotNull;
-
 public class ProductCreatorTestImpl implements ProductCreator {
 
     @Override
@@ -15,7 +12,7 @@ public class ProductCreatorTestImpl implements ProductCreator {
         String vendorId = "vendor_225674";
         String unitId = "unit_32";
         String name = "Item 1";
-        int tax = 20;
+        short tax = 20;
         String articul = "art_3305";
         String code = "code_22505";
